frontend: removed old settings files and renamed new-settings to settings
and moved files under new-settings to settings.
