--- frontends/web/src/components/badge/badge.module.css
@@ -1,21 +1,0 @@
-.container {
-  display: inline-block;
-  font-size: var(--size-small);
-  margin: 0;
-  cursor: default;
-  user-select: none;
-  -moz-user-select: none;
-  -webkit-user-select: none;
-}
-
-.generic {
-  color: var(--color-secondary);
-}
-
-.primary {
-  color: var(--color-blue);
-}
-
-.secondary {
-  color: var(--color-orange);
-}
