--- frontends/web/src/components/badge/badge.test.tsx
@@ -1,26 +1,0 @@
-/**
- * Copyright 2022 Shift Crypto AG
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { render } from '@testing-library/react';
-import { Badge } from './badge';
-
-describe('components/badge/badge', () => {
-  it('has correct class derived from its type and correct content', () => {
-    const { container } = render(<Badge type="generic">BTC</Badge>);
-    expect(container.firstChild).toHaveClass('generic');
-    expect(container).toHaveTextContent('BTC');
-  });
-});
