--- frontends/web/src/components/darkmode/darkmodetoggle.tsx
@@ -1,33 +1,0 @@
-/**
- * Copyright 2023 Shift Crypto AG
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { useTranslation } from 'react-i18next';
-import { SettingsToggle } from '../../components/settingsButton/settingsToggle';
-import { useDarkmode } from '../../hooks/darkmode';
-
-export const DarkModeToggle = () => {
-  const { t } = useTranslation();
-  const { isDarkMode, toggleDarkmode } = useDarkmode();
-
-  return (
-    <SettingsToggle
-      checked={isDarkMode}
-      id="darkMode"
-      onChange={() => toggleDarkmode(!isDarkMode)}>
-      {t('darkmode.toggle')}
-    </SettingsToggle>
-  );
-};
