+++ frontends/web/src/routes/device/bitbox02/bitbox02.tsx
@@ -24,8 +24,8 @@ import { AppUpgradeRequired } from '../../../components/appupgraderequired';
 import { CenteredContent } from '../../../components/centeredcontent/centeredcontent';
 import { Main } from '../../../components/layout';
 import { translate, TranslateProps } from '../../../decorators/translate';
-import { BB02Settings } from '../../new-settings/bb02-settings';
-import { FirmwareSetting } from '../../new-settings/components/device-settings/firmware-setting';
+import { BB02Settings } from '../../settings/bb02-settings';
+import { FirmwareSetting } from '../../settings/components/device-settings/firmware-setting';
 import { Unlock } from './unlock';
 import { Pairing } from './setup/pairing';
 import { Wait } from './setup/wait';
