+++ frontends/web/src/routes/device/bitbox02/sdcardcheck.tsx
@@ -62,7 +62,7 @@ const SDCardCheck = ({ deviceID, children }: TProps) => {
             </Button>
             <ButtonLink
               transparent
-              to={`/new-settings/device-settings/${deviceID}`}>
+              to={`/settings/device-settings/${deviceID}`}>
               {t('button.back')}
             </ButtonLink>
           </DialogButtons>
