--- frontends/web/src/routes/new-settings/type.ts
@@ -1,4 +1,0 @@
-export type TPagePropsWithSettingsTabs = {
-    deviceIDs: string[];
-    hasAccounts: boolean;
-}
