+++ frontends/web/src/routes/settings/about.tsx
@@ -19,10 +19,10 @@ import { Main, Header, GuideWrapper, GuidedContent } from '../../components/layo
 import { View, ViewContent } from '../../components/view/view';
 import { WithSettingsTabs } from './components/tabs';
 import { AppVersion } from './components/about/app-version-setting';
-import { TPagePropsWithSettingsTabs } from './type';
 import { MobileHeader } from './components/mobile-header';
 import { Guide } from '../../components/guide/guide';
 import { Entry } from '../../components/guide/entry';
+import { TPagePropsWithSettingsTabs } from './types';
 
 
 export const About = ({ deviceIDs, hasAccounts }: TPagePropsWithSettingsTabs) => {
