+++ frontends/web/src/routes/settings/components/about/app-version-setting.tsx
