+++ frontends/web/src/routes/settings/components/advanced-settings/enable-coin-control-setting.tsx
