+++ frontends/web/src/routes/settings/components/advanced-settings/enable-custom-fees-toggle-setting.tsx
