+++ frontends/web/src/routes/settings/components/appearance/activeCurrenciesDropdownSetting.tsx
