+++ frontends/web/src/routes/settings/components/appearance/defaultCurrencyDropdownSetting.tsx
