+++ frontends/web/src/routes/settings/components/appearance/displaySatsToggleSetting.tsx
