+++ frontends/web/src/routes/settings/components/device-settings/device-name-setting.tsx
