+++ frontends/web/src/routes/settings/components/device-settings/factory-reset-setting.module.css
