+++ frontends/web/src/routes/settings/components/device-settings/firmware-setting.tsx
