+++ frontends/web/src/routes/settings/components/device-settings/passphrase-setting.tsx
