+++ frontends/web/src/routes/settings/components/device-settings/show-recovery-words-setting.tsx
