+++ frontends/web/src/routes/settings/components/dropdowns/activecurrenciesdropdown.tsx
