--- frontends/web/src/routes/settings/components/fiat/fiat.module.css
@@ -1,20 +1,0 @@
-.defaultFiat {
-    color: var(--color-secondary);
-}
-
-.star {
-    background: none;
-    border: none;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    margin-left: auto;
-    margin-right: var(--space-half);
-    padding: 3px;
-}
-
-.star img {
-    color: var(--color-tertiary);
-    height: 18px;
-    width: 18px;
-}
